+++ examples/esp32/apps/memfault_demo_app/main/cmd_decl.h
@@ -8,8 +8,18 @@
 */
 #pragma once
 
+#include "freertos/FreeRTOS.h"
+#include "freertos/semphr.h"
+
 // Register system functions
-void register_system();
+void register_system(void);
 
 // Register WiFi functions
-void register_wifi();
+void register_wifi(void);
+
+// Register app-specific functions
+
+// This semaphore is used to control the Task Watchdog example task
+extern SemaphoreHandle_t g_example_task_lock;
+
+void register_app(void);