+++ examples/esp32/apps/memfault_demo_app/main/memfault_platform_config.h
@@ -7,3 +7,5 @@
 //!
 //! Platform overrides for the default configuration settings in the memfault-firmware-sdk.
 //! Default configuration settings can be found in "memfault/config.h"
+
+#define MEMFAULT_TASK_WATCHDOG_ENABLE 1