+++ scripts/mflt-build-id/src/mflt_build_id/__init__.py
@@ -46,7 +46,6 @@ if sys.version_info < (3,):
         # type: (str) -> str
         return data.encode("hex")
 
-
 else:
 
     def hexlify(data):